LocalTalk API implemented fully. Not tested. Check out synchPatientData
Any time a next page button is hit, synchPatientData should be able to
handle it. It checks that there is a valid PatientId and RecordId
(patient and record have been added to database). If not, it goes ahead
and adds them. Then it synchs the picture stored in LocalDatabase and
any Question answers that are unsynched.